bug fixes
fixed error when was deleted model from route model bind returns 404, now return component default page

+++ app/Actions/Channels/CompletlyDeleteChannelAction.php
@@ -2,28 +2,29 @@
 
 namespace App\Actions\Channels;
 
-use App\Actions\Devices\RemoveChannelFromDeviceAction;
+use App\Models\Note;
 use App\Models\Channel;
+use App\Models\Contact;
+use App\Models\RestartChannel;
 use App\Models\ChannelMulticast;
 use App\Models\ChannelQualityWithIp;
-use App\Models\Note;
-use App\Models\RestartChannel;
+use App\Actions\Devices\RemoveChannelFromDeviceAction;
 
 class CompletlyDeleteChannelAction
 {
-    public function __construct(public Channel $channel)
+    public function __construct(public object $channel)
     {
         //
     }
 
-    public function __invoke(): bool
+    public function __invoke(): void
     {
         // delete form restart channel table
         RestartChannel::where('channel_id', $this->channel->id)->delete();
-        // try {
         // delete h264 & h265
-        if (! is_null($this->channel->h264)) {
+        if (!blank($this->channel->h264)) {
             ChannelQualityWithIp::where('h264_id', $this->channel->h264->id)->delete();
+            // $this->channel->h264->ips->delete();
             $this->channel->h264->delete();
             // dispatch delete channel from device
             (new RemoveChannelFromDeviceAction(
@@ -38,8 +39,9 @@ class CompletlyDeleteChannelAction
             ))();
         }
 
-        if (! is_null($this->channel->h265)) {
+        if (!blank($this->channel->h265)) {
             ChannelQualityWithIp::where('h265_id', $this->channel->h265->id)->delete();
+            // $this->channel->h265->ips->delete();
             $this->channel->h265->delete();
             // dispatch delete channel from device
             (new RemoveChannelFromDeviceAction(
@@ -52,11 +54,11 @@ class CompletlyDeleteChannelAction
                 $this->channel->id,
                 true
             ))();
-
         }
 
-        if (! $this->channel->multicasts->isEmpty()) {
+        if (!$this->channel->multicasts->isEmpty()) {
             ChannelMulticast::where('channel_id', $this->channel->id)->delete();
+            // $this->channel->multicasts->delete();
             // dispatch delete channel from device
             (new RemoveChannelFromDeviceAction(
                 'multicast',
@@ -69,14 +71,13 @@ class CompletlyDeleteChannelAction
                 true
             ))();
         }
+        if (!$this->channel->notes->isEmpty()) {
+            Note::where('channel_id', $this->channel->id)->delete();
+        }
 
-        Note::where('channel_id', $this->channel->id)->delete();
+        // delete channel contacts
+        Contact::where('type', 'channel')->where('item_id', $this->channel->id)->delete();
 
         $this->channel->delete();
-
-        return true;
-        // } catch (\Throwable $th) {
-        //     return false;
-        // }
     }
 }
